% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inspection.R
\name{break_data}
\alias{break_data}
\title{Break up panel_data for later use}
\usage{
break_data(detail2_chr)
}
\arguments{
\item{detail2_chr}{Character representation of detail3}
}
\value{
data.frame (dplyr tbr)
}
\description{
Break up panel_data for later use
}

