% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{save.xlsx}
\alias{save.xlsx}
\title{Exporting Data to Excel Worksheet}
\usage{
save.xlsx(file, ...)
}
\arguments{
\item{xlsx}{Excel file name}
}
\description{
Provided by Rob Kabacoff
}
\examples{
save.xlsx("test.xlsx",test_object)
}

